+++ ppa/archive/models.py
@@ -5,10 +5,12 @@ from zipfile import ZipFile
 
 from cached_property import cached_property
 from django.conf import settings
+from django.core.exceptions import ValidationError
 from django.db import models
 from django.urls import reverse
 from eulxml.xmlmap import load_xmlobject_from_file
-from pairtree import pairtree_path, pairtree_client
+from pairtree import pairtree_path, pairtree_client, storage_exceptions
+import requests
 from wagtail.core.fields import RichTextField
 from wagtail.admin.edit_handlers import FieldPanel
 from wagtail.snippets.models import register_snippet
@@ -21,13 +23,50 @@ from ppa.archive.solr import PagedSolrQuery
 logger = logging.getLogger(__name__)
 
 
+#: label to use for items that are not in a collection
+NO_COLLECTION_LABEL = 'Uncategorized'
+
+
+class TrackChangesModel(models.Model):
+    ''':class:`~django.modles.Model` mixin that keeps a copy of initial
+    data in order to check if fields have been changed. Change detection
+    only works on the current instance of an object.'''
+
+    class Meta:
+        abstract = True
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        # store a copy of model data to allow for checking if
+        # it has changed
+        self.__initial = self.__dict__.copy()
+
+    def save(self, *args, **kwargs):
+        '''Saves data and reset copy of initial data.'''
+        super().save(*args, **kwargs)
+        # update copy of initial data to reflect saved state
+        self.__initial = self.__dict__.copy()
+
+    def has_changed(self, field):
+        '''check if a field has been changed'''
+        return getattr(self, field) != self.__initial[field]
+
+    def initial_value(self, field):
+        '''return the initial value for a field'''
+        return self.__initial[field]
+
+
 @register_snippet
-class Collection(models.Model):
+class Collection(TrackChangesModel):
     '''A collection of :class:`ppa.archive.models.DigitizedWork` instances.'''
     #: the name of the collection
     name = models.CharField(max_length=255)
     #: a RichText description of the collection
     description = RichTextField(blank=True)
+    #: flag to indicate collections to be excluded by default in
+    #: public search
+    exclude = models.BooleanField(default=False,
+        help_text='Exclude by default on public search.')
 
     # configure for editing in wagtail admin
     panels = [
@@ -41,24 +80,10 @@ class Collection(models.Model):
     def __str__(self):
         return self.name
 
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        # store a copy of model data to allow for checking if
-        # it has changed
-        self.__initial = self.__dict__.copy()
-
-    def save(self, *args, **kwargs):
-        """
-        Saves model and set initial state.
-        """
-        super().save(*args, **kwargs)
-        # update copy of initial data to reflect saved state
-        self.__initial = self.__dict__.copy()
-
     @property
     def name_changed(self):
         '''check if name has been changed (only works on current instance)'''
-        return self.name != self.__initial['name']
+        return self.has_changed('name')
 
     @staticmethod
     def stats():
@@ -96,24 +121,34 @@ class Collection(models.Model):
         return stats
 
 
-class DigitizedWork(models.Model, Indexable):
+class DigitizedWork(TrackChangesModel, Indexable):
     '''
     Record to manage digitized works included in PPA and store their basic
     metadata.
     '''
-    # stub record to manage digitized works included in PPA
-    # basic metadata
-    # - title, author, place of publication, date
-    # added, updated
-    # original id / url?
+    HATHI = 'HT'
+    OTHER = 'O'
+    SOURCE_CHOICES = (
+        (HATHI, 'HathiTrust'),
+        (OTHER, 'Other'),
+    )
+    #: source of the record, HathiTrust or elsewhere
+    source = models.CharField(
+        max_length=2, choices=SOURCE_CHOICES, default=HATHI,
+        help_text='Source of the record.')
     #: source identifier; hathi id for HathiTrust materials
-    source_id = models.CharField(max_length=255, unique=True)
+    source_id = models.CharField(
+        max_length=255, unique=True, verbose_name='Source ID',
+        help_text='Source identifier. Unique identifier without spaces; ' +
+                  'used for site URL. (HT id for HathiTrust materials.)')
     #: source url where the original can be accessed
-    source_url = models.URLField(max_length=255)
+    source_url = models.URLField(
+        max_length=255, verbose_name='Source URL', blank=True,
+        help_text='URL where the source item can be accessed')
     #: record id; for Hathi materials, used for different copies of
     #: the same work or for different editions/volumes of a work
     record_id = models.CharField(
-        max_length=255,
+        max_length=255, blank=True,
         help_text='For HathiTrust materials, record id (use to aggregate ' + \
                   'copies or volumes).')
     #: title of the work; using TextField to allow for long titles
@@ -122,15 +157,18 @@ class DigitizedWork(models.Model, Indexable):
     subtitle = models.TextField(blank=True, default='',
                                 help_text='Subtitle, if any (optional)')
     #: sort title: title without leading non-sort characters, from marc
-    sort_title = models.TextField(default='',
-                                  help_text='Sort title from MARC record')
+    sort_title = models.TextField(
+        default='',
+        help_text='Sort title from MARC record or title without leading article')
     #: enumeration/chronology (hathi-specific)
     enumcron = models.CharField('Enumeration/Chronology', max_length=255,
                                 blank=True)
     # TODO: what is the generic/non-hathi name for this? volume/version?
 
     # NOTE: may eventually to convert to foreign key
-    author = models.CharField(max_length=255, blank=True)
+    author = models.CharField(
+        max_length=255, blank=True,
+        help_text='Authorized name of the author, last name first.')
     #: place of publication
     pub_place = models.CharField('Place of Publication', max_length=255,
                                  blank=True)
@@ -153,6 +191,19 @@ class DigitizedWork(models.Model, Indexable):
     #: date of last modification of the local record
     updated = models.DateTimeField(auto_now=True)
 
+    PUBLIC = 'P'
+    SUPPRESSED = 'S'
+    STATUS_CHOICES = (
+        (PUBLIC, 'Public'),
+        (SUPPRESSED, 'Suppressed'),
+    )
+    #: status of record; currently choices are public or suppressed
+    status = models.CharField(
+        max_length=2, choices=STATUS_CHOICES, default=PUBLIC,
+        help_text='Changing status to suppressed will remove rsync data ' +
+        'for that volume and remove from the public index. This is ' +
+        'currently not reversible; use with caution.')
+
     class Meta:
         ordering = ('sort_title',)
 
@@ -168,16 +219,9 @@ class DigitizedWork(models.Model, Indexable):
         return self.source_id
 
     @property
-    def srcid(self):
-        '''alias for :attr:`source_id` for consistency with solr attributes'''
-        # hopefully temporary workaround until solr fields made consistent
-        return self.source_id
-
-    @property
-    def src_url(self):
-        '''alias for :attr:`source_url` for consistency with solr attributes'''
-        # hopefully temporary workaround until solr fields made consistent
-        return self.source_url
+    def is_suppressed(self):
+        '''Item has been suppressed (based on :attr:`status`).'''
+        return self.status == self.SUPPRESSED
 
     def display_title(self):
         '''admin display title to allow displaying title but sorting on sort_title'''
@@ -186,10 +230,48 @@ class DigitizedWork(models.Model, Indexable):
     display_title.admin_order_field = 'sort_title'
     display_title.allow_tags = True
 
+    def is_public(self):
+        '''admin display field indicating if record is public or suppressed'''
+        return self.status == self.PUBLIC
+    is_public.short_description = 'Public'
+    is_public.boolean = True
+    is_public.admin_order_field = 'status'
+
     #: regular expresion for cleaning preliminary text from publisher names
     printed_by_re = r'^(Printed)?( and )?(Pub(.|lished|lisht)?)?( and sold)? (by|for|at)( the)? ?'
     # Printed by/for (the); Printed and sold by; Printed and published by;
     # Pub./Published/Publisht at/by/for the
+
+    def save(self, *args, **kwargs):
+        # if status has changed so that object is now suppressed and this
+        # is a HathiTrust item, remove pairtree data
+        if self.has_changed('status') and self.status == self.SUPPRESSED \
+          and self.source == DigitizedWork.HATHI:
+            self.delete_hathi_pairtree_data()
+
+        # source id is used as Solr identifier; if it changes, remove
+        # the old record from Solr before saving with the new identifier
+        # NOTE: source id edit only supported for non-hathi content; should
+        # be prevented by validation in clean method
+        if self.has_changed('source_id'):
+            new_source_id = self.source_id
+            self.source_id = self.initial_value('source_id')
+            self.remove_from_index(params={"commitWithin": 3000})
+            self.source_id = new_source_id
+
+        super().save(*args, **kwargs)
+
+    def clean(self):
+        '''Add custom validation to trigger a save error in the admin
+        if someone tries to unsuppress a record that has been suppressed
+        (not yet supported).'''
+        if self.has_changed('status') and self.status != self.SUPPRESSED:
+            raise ValidationError('Unsuppressing records not yet supported.')
+
+        # should not be editable in admin, but add a validation check
+        # just in case
+        if self.has_changed('source_id') and self.source == self.HATHI:
+            raise ValidationError('Changing source ID for HathiTrust records is not supported')
 
     def populate_from_bibdata(self, bibdata):
         '''Update record fields based on Hathi bibdata information.
@@ -349,10 +431,17 @@ class DigitizedWork(models.Model, Indexable):
 
     def index_data(self):
         '''data for indexing in Solr'''
+
+        # When an item has been suppressed, return id only.
+        # This will blank out any previously indexed values, and item
+        # will not be findable by any public searchable fields.
+        if self.status == self.SUPPRESSED:
+            return {'id': self.source_id}
+
         return {
             'id': self.source_id,
-            'srcid': self.source_id,
-            'src_url': self.source_url,
+            'source_id': self.source_id,
+            'source_url': self.source_url,
             'title': self.title,
             'subtitle': self.subtitle,
             'sort_title': self.sort_title,
@@ -361,12 +450,14 @@ class DigitizedWork(models.Model, Indexable):
             'publisher': self.publisher,
             'enumcron': self.enumcron,
             'author': self.author,
-            'collections': [collection.name for collection
-                            in self.collections.all()],
-            # general purpose multivalued field, currently only
-            # includes public notes in this method, other fields
-            # copied in Solr schema.
-            'text': [self.public_notes],
+            # set default value to simplify queries to find uncollected items
+            # (not set in Solr schema because needs to be works only)
+            'collections':
+                [collection.name for collection in self.collections.all()]
+                if self.collections.exists()
+                else [NO_COLLECTION_LABEL],
+            # public notes field for display on site_name
+            'notes': self.public_notes,
             # hard-coded to distinguish from & sort with pages
             'item_type': 'work',
             'order': '0',
@@ -391,17 +482,39 @@ class DigitizedWork(models.Model, Indexable):
         # pairtree encoded version of the id
         return pairtree_path.id_encode(self.hathi_pairtree_id)
 
+    def hathi_pairtree_client(self):
+        '''Initialize a pairtree client for the pairtree datastore this
+         object belongs to, based on its Hathi prefix id.'''
+        return pairtree_client.PairtreeStorageClient(
+            self.hathi_prefix,
+            os.path.join(settings.HATHI_DATA, self.hathi_prefix))
+
     def hathi_pairtree_object(self, ptree_client=None):
-        '''get a pairtree object for the current work'''
+        '''get a pairtree object for the current work
+
+        :param ptree_client: optional
+            :class:`pairtree_client.PairtreeStorageClient` if one has
+            already been initialized, to avoid repeated initialization
+            (currently used in hathi_import manage command)
+        '''
         if ptree_client is None:
             # get pairtree client if not passed in
-            ptree_client = pairtree_client.PairtreeStorageClient(
-                self.hathi_prefix,
-                os.path.join(settings.HATHI_DATA, self.hathi_prefix))
+            ptree_client = self.hathi_pairtree_client()
 
         # return the pairtree object for current work
         return ptree_client.get_object(self.hathi_pairtree_id,
                                        create_if_doesnt_exist=False)
+
+    def delete_hathi_pairtree_data(self):
+        '''Delete pairtree object from the pairtree datastore.'''
+        logger.info('Deleting pairtree data for %s', self.source_id)
+        try:
+            self.hathi_pairtree_client() \
+                .delete_object(self.hathi_pairtree_id)
+        except storage_exceptions.ObjectNotFoundException:
+            # data is already gone; warn, but not an error
+            logger.warn('Pairtree deletion failed; object not found %s',
+                        self.source_id)
 
     def _hathi_content_path(self, ext, ptree_client=None):
         '''path to zipfile within the hathi contents for this work'''
@@ -428,9 +541,19 @@ class DigitizedWork(models.Model, Indexable):
         '''Get page content for this work from Hathi pairtree and return
         data to be indexed in solr.'''
 
+        # If an item has been suppressed or is from a source other than
+        # hathi, bail out. No pages to index.
+        if self.is_suppressed or self.source != self.HATHI:
+            return
+
         # load mets record to pull metadata about the images
-        mmets = load_xmlobject_from_file(self.hathi_metsfile_path(),
-                                         MinimalMETS)
+        try:
+            mmets = load_xmlobject_from_file(self.hathi_metsfile_path(),
+                                             MinimalMETS)
+        except storage_exceptions.ObjectNotFoundException:
+            logger.error('Pairtree data for %s not found but status is %s',
+                         self.source_id, self.get_status_display())
+            return
 
         # read zipfile contents in place, without unzipping
         with ZipFile(self.hathi_zipfile_path()) as ht_zip:
@@ -444,7 +567,7 @@ class DigitizedWork(models.Model, Indexable):
                     try:
                         yield {
                             'id': '%s.%s' % (self.source_id, page.text_file.sequence),
-                            'srcid': self.source_id,   # for grouping with work record
+                            'source_id': self.source_id,   # for grouping with work record
                             'content': pagefile.read().decode('utf-8'),
                             'order': page.order,
                             'label': page.display_label,
@@ -457,13 +580,18 @@ class DigitizedWork(models.Model, Indexable):
     def get_metadata(self, metadata_format):
         '''Get metadata for this item in the specified format.
         Currently only supports marc.'''
-
         if metadata_format == 'marc':
             # get metadata from hathi bib api and serialize
             # as binary marc
-            bib_api = HathiBibliographicAPI()
-            bibdata = bib_api.record('htid', self.source_id)
-            return bibdata.marcxml.as_marc()
+            if self.source == DigitizedWork.HATHI:
+                bib_api = HathiBibliographicAPI()
+                bibdata = bib_api.record('htid', self.source_id)
+                return bibdata.marcxml.as_marc()
+
+            # TBD: can we get MARC records from oclc?
+            # or should we generate dublin core from db metadata?
+
+            return ''
 
         # error for unknown
         raise ValueError('Unsupported format %s' % metadata_format)