+++ ppa/editorial/__init__.py
@@ -1,5 +1,7 @@
 '''
-Django app for editorial content
+Management and display of editorial content. Includes custom page models
+for an editorial list page and editorial content pages, structured
+roughly like a scholarly blog..
 '''
 
 default_app_config = 'ppa.editorial.apps.EditorialConfig'