+++ sitemedia/js/search.js
@@ -20,33 +20,27 @@ $(function(){
     const $$collectionInputs = $('#collections input')
     const $$textInputs = $('input[type="text"]')
     const $$relevanceOption = $('#sort .item[data-value="relevance"]')
-    const $$advancedSearchButton = $('.show-advanced')
+    const $$advancedSearchButton = $('.show-advanced button')
 
     /* bindings */
     archiveSearchForm.onStateChange(submitForm)
     $$clearDatesLink.click(onClearDates)
-    $$collectionInputs.change(onCollectionChange)
     $$advancedSearchButton.click(toggleAdvancedSearch)
+    $$textInputs.keyup(onTextInputChange)
     onPageLoad() // misc functions that run once on page load
 
     $$collectionInputs
         .focus(e => $(e.target).parent().addClass('focus')) // make collection buttons focusable
-        .blur(e => $(e.target).parent().removeClass('focus')) 
+        .blur(e => $(e.target).parent().removeClass('focus'))
+        .change(e => $(e.target).parent().toggleClass('active'))
         .keypress(e => { if (e.which == 13) $(e.target).click() }) // pressing enter "clicks" them
-    
+
     /* functions */
     function submitForm(state) {
         if (!validate()) return // don't submit an invalid form
         state = state.filter(field => field.value != '') // filter out empty fields
-        if (state.filter(field => $$textInputs.get().map(el => el.name).includes(field.name)).length == 0) { // if no text query,
-            $$relevanceOption.addClass('disabled') // disable relevance
-            let sort = state.find(field => field.name == 'sort') // check if a sort was set
-            if (sort && sort.value == 'relevance') { // and if it was relevance,
-                $$sortDropdown.dropdown('set selected', 'title_asc') // set to title A-Z
-            }
-        }
-        else {
-            $$relevanceOption.removeClass('disabled') // enable relevance sort
+        if (state.filter(field => field.name == 'collections').length == 0) { // if the user manually turned off all collections...
+            state.push({ name: "collections", value: "" }) // add a blank value to indicate that specific case
         }
         let url = `?${$.param(state)}` // serialize state using $.param to make querystring
         window.history.pushState(state, 'PPA Archive Search', url) // update the URL bar
@@ -55,6 +49,7 @@ $(function(){
                 'X-Requested-With': 'XMLHttpRequest'
             }
         })
+        $('.workscount').addClass('loading') // turn on the loader
         req.then(res => res.text()).then(html => { // submit the form and get html back
             $$paginationTop.html($(html).find('.page-controls').html()) // update the top pagination
             dateHistogram.update(JSON.parse($(html).find('pre.facets').html())) // update the histogram
@@ -64,11 +59,19 @@ $(function(){
                 bubbles: true,
                 cancelable: true
             }))
+            $('.workscount').removeClass('loading') // turn off the loader
         })
+        advancedSearchIndicator()
     }
 
     function validate() {
         if (!$$minDateInput[0].checkValidity() || !$$maxDateInput[0].checkValidity()) {
+            $('.validation').css('visibility', 'visible')
+            return false
+        }
+        // validate that min occurs before max when both are set
+        if ($$minDateInput.val() && $$maxDateInput.val() &&
+            $$minDateInput.val() > $$maxDateInput.val()) {
             $('.validation').css('visibility', 'visible')
             return false
         }
@@ -83,11 +86,30 @@ $(function(){
         $$maxDateInput[0].dispatchEvent(new Event('input'))
     }
 
-    function onCollectionChange(event) {
-        $(event.currentTarget).parent().toggleClass('active')
+    function onTextInputChange(event) {
+        // update sort options based on changes to text input fields
+
+        // if any text inputs now have content
+        if ($$textInputs.get().filter(el => $.trim($(el).val()) != '').length) {
+            // enable and select relevance sort
+            $$relevanceOption.removeClass('disabled')
+            $$sortDropdown.dropdown('set selected', 'relevance')
+
+        // no text inputs have content now
+        } else {
+            // disable relevance sort option
+            $$relevanceOption.addClass('disabled')
+            // if relevance sort was selected, set back to title
+            let sort = archiveSearchForm.state.find(field => field.name == 'sort')
+            if (sort && sort.value == 'relevance') {
+                $$sortDropdown.dropdown('set selected', 'title_asc')
+            }
+        }
     }
 
     function onPageLoad() {
+        // remove no-js styles for CSS since Javascript is running
+        $('.form').removeClass('no-js')
         dateHistogram.update(JSON.parse($('.ajax-container pre.facets').html())) // render the histogram initially
         $$collectionInputs.filter(':disabled').parent().addClass('disabled') // disable empty collections
         $('.question-popup').popup() // initialize the question popup
@@ -97,9 +119,46 @@ $(function(){
         $('.form').keydown(e => { if (e.which === 13) e.preventDefault() }) // don't allow enter key to submit the search
         $$textInputs.each((_, el) => clearable(el)) // make text inputs clearable
         validate()
+        if (sessionStorage.getItem('ppa-adv-search') == 'open') { // open advanced search without animating it
+            $('.show-advanced').addClass('active')
+            $('.advanced.segment').css('display', 'flex')
+            $('.advanced.column').css('display', 'inline-block')
+        }
+        advancedSearchIndicator()
+    }
+
+    function advancedSearchOn() {
+        $('.show-advanced').addClass('active')
+        $('.advanced.segment').css('display', 'flex') // if we don't manually set flex here, jQuery can't infer it
+        $('.advanced.column').css('display', 'inline-block') // column shouldn't be flex
+        $('.advanced').hide().slideDown() // hide sets display to none while animating
+        sessionStorage.setItem('ppa-adv-search', 'open') // remember the value for this session
+    }
+
+    function advancedSearchOff() {
+        $('.show-advanced').removeClass('active')
+        $('.advanced').slideUp()
+        sessionStorage.setItem('ppa-adv-search', 'closed')
     }
 
     function toggleAdvancedSearch() {
-        $('.advanced').slideToggle()
+        $('.advanced').is(':hidden') ? advancedSearchOn() : advancedSearchOff()
     }
+
+    function advancedSearchIndicator() { // if any adv. search fields are active, show it
+        if ($('.advanced input').get().map(e => e.value).find(v => v != '')) {
+            $('.show-advanced .search-active').fadeIn()
+        }
+        else {
+            $('.show-advanced .search-active').fadeOut()
+        }
+    }
+
+    bodymovin.loadAnimation({ // set up the loader animation
+        container: document.getElementById('bm'),
+        renderer: 'svg',
+        loop: true,
+        autoplay: true,
+        path: '/static/img/loader/searchLoading.json'
+    })
 })