+++ CHANGELOG.rst
@@ -3,6 +3,49 @@
 CHANGELOG
 =========
 
+3.0 - Initial public version (soft launch)
+------------------------------------------
+
+**PPA 3.0 is a completely new implementation of the Princeton Prosody
+Archive project. The 3.0 is used here for what would normally be a 1.0 release
+as a way to credit and differentiate from previous versions of PPA.**
+
+Admin & data curation functionality
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+* As an admin, I want to manually enter bibliographic information into editable fields so that users can view and search citations for works not available in Hathi.
+* As an admin, I want to suppress items from the site so that I can pull content that should not be included or was wrongly added as I am going through and assigning collections to archive volumes.
+
+
+Search improvements
+^^^^^^^^^^^^^^^^^^^
+* As a user, I want keyword searches to prioritize matches in the author, title and public notes fields so that I can easily find works using keywords.
+* As a user, I should not see suppressed items in search results or item display so that my results are not cluttered by items not meant to be part of the archive.
+* As a user, I want to exclude or include items from any of the collections in PPA so that I can refine my search to include relevant items.
+* As a user, I want the advanced search pulldown state that I have selected to be preserved when I reload the page so that my view of the search form is stable and consistent without having to continually modify my selection.
+* Automatically change default sort to Relevance for keyword searches
+* Change `srcid` to `source_id` for fielded search
+* bugfix: non-sequential publication dates break search validation
+* bugfix: Using actual numbers for date range causes works without
+  a date to go missing when form is submitted
+
+Content management
+^^^^^^^^^^^^^^^^^^
+* As a content editor, I want to be able to add and order multiple authors to an editorial so I can correctly attribute work.
+* As a content editor, I want to list people who contributed to the project so that I can give credit to everyone who was involved in it.
+
+UI/UX/Design updates
+^^^^^^^^^^^^^^^^^^^^
+* Refinements to the search form
+  - collapsible advanced search, hidden by default
+  - visual indicator if filters are active in the advanced search
+  - revised styles for collection filters
+* Indicator for search in-progress
+* Add a "jump to top" button on search results
+* Styles for editorial list page, editorial post including image captions
+  and footnotes
+* Updated error pages
+
+
 0.11.0
 ------